+++ src/__init__.py
@@ -1,14 +1,1 @@
-
-# SPDX-License-Identifier: MIT
-# --------------------------------------------------------------
-#|																|
-#|             Copyright 2023 - 2023, Amulya Paritosh			|
-#|																|
-#|  This file is part of Component Library Plugin for FreeCAD.	|
-#|																|
-#|               This file was created as a part of				|
-#|              Google Summer Of Code Program - 2023			|
-#|																|
-# --------------------------------------------------------------
-
-from .app.app import create_app
+from .app import create_app
