+++ src/config.py
@@ -1,19 +1,16 @@
-
 # SPDX-License-Identifier: MIT
 # --------------------------------------------------------------
-#|																|
-#|             Copyright 2023 - 2023, Amulya Paritosh			|
-#|																|
-#|  This file is part of Component Library Plugin for FreeCAD.	|
-#|																|
-#|               This file was created as a part of				|
-#|              Google Summer Of Code Program - 2023			|
-#|																|
+# |																|
+# |             Copyright 2023 - 2023, Amulya Paritosh			|
+# |																|
+# |  This file is part of Component Library Plugin for FreeCAD.	|
+# |																|
+# |               This file was created as a part of				|
+# |              Google Summer Of Code Program - 2023			|
+# |																|
 # --------------------------------------------------------------
 
 import os
-import sys
-import logging
 
 basedir: str = os.path.abspath(os.path.dirname(__file__))
 
@@ -45,5 +42,5 @@ class Config:
     DEBUG = True
     TESTING = True
 
-    SQLALCHEMY_DATABASE_URI: str = f"sqlite:///{basedir}/app.db"
+    SQLALCHEMY_DATABASE_URI: str = f"sqlite:///{basedir}/app/app.db"
     SQLALCHEMY_TRACK_MODIFICATIONS = False
