+++ src/models/files/__init__.py
@@ -1,3 +1,2 @@
 from .models import File, FileType
-from .operations import upload_to_github
 from .schemas import FileSchema, files_schema
