+++ src/models/licenses/schemas.py
@@ -1,20 +1,19 @@
-
 # SPDX-License-Identifier: MIT
 # --------------------------------------------------------------
-#|																|
-#|             Copyright 2023 - 2023, Amulya Paritosh			|
-#|																|
-#|  This file is part of Component Library Plugin for FreeCAD.	|
-#|																|
-#|               This file was created as a part of				|
-#|              Google Summer Of Code Program - 2023			|
-#|																|
+# |																|
+# |             Copyright 2023 - 2023, Amulya Paritosh			|
+# |																|
+# |  This file is part of Component Library Plugin for FreeCAD.	|
+# |																|
+# |               This file was created as a part of				|
+# |              Google Summer Of Code Program - 2023			|
+# |																|
 # --------------------------------------------------------------
 
 from flask_sqlalchemy.session import Session
 from sqlalchemy.orm.scoping import scoped_session
 
-from ..database import db, ma
+from ...database import db, ma
 from .models import SPDX
 
 
@@ -33,10 +32,12 @@ class SPDXSchema(ma.SQLAlchemyAutoSchema):
         print(result)
         ```
     """
+
     class Meta:
         model = SPDX
         load_instance = True
         sqla_session: scoped_session[Session] = db.session
+
 
 spdx_schema = SPDXSchema()
 spdxs_schema = SPDXSchema(many=True)
