+++ src/models/metadatas/schemas.py
@@ -1,21 +1,20 @@
-
 # SPDX-License-Identifier: MIT
 # --------------------------------------------------------------
-#|																|
-#|             Copyright 2023 - 2023, Amulya Paritosh			|
-#|																|
-#|  This file is part of Component Library Plugin for FreeCAD.	|
-#|																|
-#|               This file was created as a part of				|
-#|              Google Summer Of Code Program - 2023			|
-#|																|
+# |																|
+# |             Copyright 2023 - 2023, Amulya Paritosh			|
+# |																|
+# |  This file is part of Component Library Plugin for FreeCAD.	|
+# |																|
+# |               This file was created as a part of				|
+# |              Google Summer Of Code Program - 2023			|
+# |																|
 # --------------------------------------------------------------
 
 from flask_sqlalchemy.session import Session
 from marshmallow import fields
 from sqlalchemy.orm.scoping import scoped_session
 
-from ..database import db, ma
+from ...database import db, ma
 from .models import Metadata
 
 
@@ -34,10 +33,12 @@ class MetadataSchema(ma.SQLAlchemyAutoSchema):
     """
 
     license_id = fields.String()
+
     class Meta:
         model = Metadata
         load_instance = True
         sqla_session: scoped_session[Session] = db.session
+
 
 metadata_schema = MetadataSchema()
 metadatas_schema = MetadataSchema(many=True)
