+++ src/validation/exceptions.py
@@ -1,0 +1,14 @@
+class InvalidEmail(Exception):
+    """
+    Exception raised for invalid email addresses.
+
+    This exception is raised when an email address is determined to be invalid.
+    """
+
+
+class InvalidURL(Exception):
+    """
+    Exception raised for invalid URLs.
+
+    This exception is raised when a URL is determined to be invalid.
+    """
