+++ src/validation/validators.py
@@ -1,34 +1,19 @@
-
 # SPDX-License-Identifier: MIT
 # --------------------------------------------------------------
-#|																|
-#|             Copyright 2023 - 2023, Amulya Paritosh			|
-#|																|
-#|  This file is part of Component Library Plugin for FreeCAD.	|
-#|																|
-#|               This file was created as a part of				|
-#|              Google Summer Of Code Program - 2023			|
-#|																|
+# |																|
+# |             Copyright 2023 - 2023, Amulya Paritosh			|
+# |																|
+# |  This file is part of Component Library Plugin for FreeCAD.	|
+# |																|
+# |               This file was created as a part of				|
+# |              Google Summer Of Code Program - 2023			|
+# |																|
 # --------------------------------------------------------------
 
 import re
 from urllib.parse import urlparse
 
-
-class InvalidEmail(Exception):
-    """
-    Exception raised for invalid email addresses.
-
-    This exception is raised when an email address is determined to be invalid.
-    """
-
-class InvalidURL(Exception):
-    """
-    Exception raised for invalid URLs.
-
-    This exception is raised when a URL is determined to be invalid.
-    """
-
+from .exceptions import InvalidEmail, InvalidURL
 
 
 def email_validator(email: str) -> str | None:
@@ -57,6 +42,7 @@ def email_validator(email: str) -> str | None:
     if not re.match(r"[^@]+@[^@]+\.[^@]+", email):
         raise InvalidEmail()
     return email
+
 
 def url_validator(url: str) -> str | None:
     """
